+++ crates/hdk/src/hdk.rs
@@ -69,7 +69,6 @@ pub trait HdkT: Send + Sync {
     ) -> ExternResult<Vec<LinkDetails>>;
     // P2P
     fn call(&self, call: Vec<Call>) -> ExternResult<Vec<ZomeCallResponse>>;
-    fn call_remote(&self, call_remote: Vec<CallRemote>) -> ExternResult<Vec<ZomeCallResponse>>;
     fn emit_signal(&self, app_signal: AppSignal) -> ExternResult<()>;
     fn remote_signal(&self, remote_signal: RemoteSignal) -> ExternResult<()>;
     // Random
@@ -188,9 +187,6 @@ impl HdkT for ErrHdk {
     }
     // P2P
     fn call(&self, _: Vec<Call>) -> ExternResult<Vec<ZomeCallResponse>> {
-        Self::err()
-    }
-    fn call_remote(&self, _: Vec<CallRemote>) -> ExternResult<Vec<ZomeCallResponse>> {
         Self::err()
     }
     fn emit_signal(&self, _: AppSignal) -> ExternResult<()> {
@@ -355,9 +351,6 @@ impl HdkT for HostHdk {
     }
     fn call(&self, call: Vec<Call>) -> ExternResult<Vec<ZomeCallResponse>> {
         host_call::<Vec<Call>, Vec<ZomeCallResponse>>(__call, call)
-    }
-    fn call_remote(&self, call_remote: Vec<CallRemote>) -> ExternResult<Vec<ZomeCallResponse>> {
-        host_call::<Vec<CallRemote>, Vec<ZomeCallResponse>>(__call_remote, call_remote)
     }
     fn emit_signal(&self, app_signal: AppSignal) -> ExternResult<()> {
         host_call::<AppSignal, ()>(__emit_signal, app_signal)
