--- crates/holochain/src/core/ribosome/host_fn/call_remote.rs
@@ -1,198 +1,0 @@
-use crate::core::ribosome::CallContext;
-use crate::core::ribosome::HostFnAccess;
-use crate::core::ribosome::RibosomeT;
-use futures::future::join_all;
-use holochain_p2p::HolochainP2pDnaT;
-use holochain_types::prelude::*;
-use holochain_wasmer_host::prelude::WasmError;
-use crate::core::ribosome::RibosomeError;
-use std::sync::Arc;
-
-pub fn call_remote(
-    _ribosome: Arc<impl RibosomeT>,
-    call_context: Arc<CallContext>,
-    inputs: Vec<CallRemote>,
-) -> Result<Vec<ZomeCallResponse>, WasmError> {
-    match HostFnAccess::from(&call_context.host_context()) {
-        HostFnAccess {
-            write_network: Permission::Allow,
-            agent_info: Permission::Allow,
-            ..
-        } => {
-            let from_agent = super::agent_info::agent_info(_ribosome, call_context.clone(), ())?
-                .agent_latest_pubkey;
-            // it is the network's responsibility to handle timeouts and return an Err result in that case
-            let results: Vec<Result<SerializedBytes, _>> =
-                tokio_helper::block_forever_on(async move {
-                    join_all(inputs.into_iter().map(|input| async {
-                        let CallRemote {
-                            target_agent,
-                            zome_name,
-                            fn_name,
-                            cap_secret,
-                            payload,
-                        } = input;
-                        call_context
-                            .host_context()
-                            .network()
-                            .call_remote(from_agent.clone(), target_agent, zome_name, fn_name, cap_secret, payload)
-                            .await
-                    }))
-                    .await
-                });
-            let results: Result<Vec<_>, _> = results
-                .into_iter()
-                .map(|result| match result {
-                    Ok(r) => match ZomeCallResponse::try_from(r) {
-                        Ok(v) => Ok(v),
-                        Err(e) => Err(e),
-                    },
-                    Err(e) => Ok(ZomeCallResponse::NetworkError(e.to_string())),
-                })
-                .collect();
-
-            Ok(results?)
-        }
-        _ => Err(WasmError::Host(RibosomeError::HostFnPermissions(
-            call_context.zome.zome_name().clone(),
-            call_context.function_name().clone(),
-            "call_remote".into()
-        ).to_string()))
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "slow_tests")]
-pub mod wasm_test {
-    use crate::conductor::api::ZomeCall;
-    use crate::conductor::interface::websocket::test_utils::setup_app;
-    use crate::core::ribosome::ZomeCallResponse;
-    use hdk::prelude::*;
-    use holochain_types::prelude::*;
-    use holochain_types::test_utils::fake_agent_pubkey_1;
-    use holochain_types::test_utils::fake_agent_pubkey_2;
-    use holochain_wasm_test_utils::TestWasm;
-    pub use holochain_zome_types::capability::CapSecret;
-    use holochain_zome_types::cell::CellId;
-    use holochain_zome_types::ExternIO;
-
-    #[tokio::test(flavor = "multi_thread")]
-    /// we can call a fn on a remote
-    async fn call_remote_test() {
-        // ////////////
-        // START DNA
-        // ////////////
-
-        let dna_def = DnaDef {
-            name: "call_remote_test".to_string(),
-            uid: "ba1d046d-ce29-4778-914b-47e6010d2faf".to_string(),
-            properties: SerializedBytes::try_from(()).unwrap(),
-            zomes: vec![TestWasm::WhoAmI.into()].into(),
-        };
-        let dna_file = DnaFile::new(dna_def, vec![TestWasm::WhoAmI.into()])
-            .await
-            .unwrap();
-
-        // //////////
-        // END DNA
-        // //////////
-
-        // ///////////
-        // START ALICE
-        // ///////////
-
-        let alice_agent_id = fake_agent_pubkey_1();
-        let alice_cell_id = CellId::new(dna_file.dna_hash().to_owned(), alice_agent_id.clone());
-        let alice_installed_cell = InstalledCell::new(alice_cell_id.clone(), "alice_handle".into());
-
-        // /////////
-        // END ALICE
-        // /////////
-
-        // /////////
-        // START BOB
-        // /////////
-
-        let bob_agent_id = fake_agent_pubkey_2();
-        let bob_cell_id = CellId::new(dna_file.dna_hash().to_owned(), bob_agent_id.clone());
-        let bob_installed_cell = InstalledCell::new(bob_cell_id.clone(), "bob_handle".into());
-
-        // ///////
-        // END BOB
-        // ///////
-
-        // ///////////////
-        // START CONDUCTOR
-        // ///////////////
-
-        let mut dna_store = MockDnaStore::new();
-
-        dna_store.expect_get().return_const(Some(dna_file.clone()));
-        dna_store
-            .expect_add_dnas::<Vec<_>>()
-            .times(2)
-            .return_const(());
-        dna_store
-            .expect_add_entry_defs::<Vec<_>>()
-            .times(2)
-            .return_const(());
-
-        let (_tmpdir, _app_api, handle) = setup_app(
-            vec![(alice_installed_cell, None), (bob_installed_cell, None)],
-            dna_store,
-        )
-        .await;
-
-        // /////////////
-        // END CONDUCTOR
-        // /////////////
-
-        // BOB INIT (to do cap grant)
-
-        let _ = handle
-            .call_zome(ZomeCall {
-                cell_id: bob_cell_id,
-                zome_name: TestWasm::WhoAmI.into(),
-                cap_secret: None,
-                fn_name: "set_access".into(),
-                payload: ExternIO::encode(()).unwrap(),
-                provenance: bob_agent_id.clone(),
-            })
-            .await
-            .unwrap();
-
-        // ALICE DOING A CALL
-
-        let output = handle
-            .call_zome(ZomeCall {
-                cell_id: alice_cell_id,
-                zome_name: TestWasm::WhoAmI.into(),
-                cap_secret: None,
-                fn_name: "whoarethey".into(),
-                payload: ExternIO::encode(&bob_agent_id).unwrap(),
-                provenance: alice_agent_id,
-            })
-            .await
-            .unwrap()
-            .unwrap();
-
-        match output {
-            ZomeCallResponse::Ok(guest_output) => {
-                let agent_info: AgentInfo = guest_output.decode().unwrap();
-                assert_eq!(
-                    agent_info.agent_initial_pubkey,
-                    bob_agent_id
-                );
-                assert_eq!(
-                    agent_info.agent_latest_pubkey,
-                    bob_agent_id
-                );
-            }
-            _ => unreachable!(),
-        }
-
-        let shutdown = handle.take_shutdown_handle().unwrap();
-        handle.shutdown();
-        shutdown.await.unwrap().unwrap();
-    }
-}
